fix: revert simplification that removed patch splitting
The splitting lead to cases in which a lot of changes part of a hunk could not be applied due to a single failure. This is exactly what we argued and showed in the synchronization study. I do not remember why I had decided to remove the splitting in the first place.